Renamed and moved
Renamed and moved classes, to match with the java and bukkit conventions
